Add files via upload
Two part case study of office building project and comparison of it's energy consumption, with the first part being analysis in 3 different construction zones, in the cities of Sidney, Belgrade and Yakutsk. The second part of the study is analysis of building sensitivity in the same zone but with three different types of wall, without insolation, one-side insolated wall and wall with insolation on both sides.