Add files via upload
This is a repeated upload of old folder with updated data about thermal resistance data for each wall. Only the final values were added. Also as we talked, one more member joined our group, Rimi Janka 918801, and we added his name into the folder of the group but he is not in that survey that we had to fill few weeks ago. We hope this won't cause much trouble.